Input Management and GfxEngine
Basic Input Management with SDL implemented and the GfxEngine Skeleton is running along with the engine. Gave it the Singleton pattern as well, as control over the graphics may be separated from the engine, but it allows for control over the graphics separate from the engine itself. Will see how this works out